requirements.txt
alembic==1.11.1
aniso8601==9.0.1
apispec==6.3.0
async-timeout==4.0.2
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==40.0.2
dynaconf==3.1.12
ecdsa==0.18.0
fabric==3.0.1
Faker==18.9.0
Flask==2.3.2
flask-apispec==0.11.4
Flask-Caching==2.0.2
Flask-JSON==0.4.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
flask-rest-paginate==1.1.3
Flask-RESTful==0.3.10
Flask-Script==2.0.6
flask-serialize==2.1.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.6.0
invoke==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
paramiko==3.1.0
permissive-dict==1.0.4
psycopg2
pyasn1==0.5.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
redis==4.5.5
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
urllib3==1.26.16
webargs==8.2.0
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
